리펙토링 : float 자료형 double 자료형으로 변경

+++ src/main/java/com/devrun/dto/MylectureReviewDTO.java
@@ -9,5 +9,5 @@ public class MylectureReviewDTO {
 	private String userId;
 	private String reviewContent;
 	private Date reviewDate;
-	private float reviewRating;
+	private double reviewRating;
 }
